+++ src/utils/time.ts
@@ -1,5 +1,5 @@
 export const prettyTime = (seconds: number): string => {
-  const nSeconds = seconds % 60;
+  const nSeconds = Math.floor(seconds) % 60;
   let nMinutes = Math.floor(seconds / 60);
   const nHours = Math.floor(nMinutes / 60);